Anatree

Description

The Tree type represents a dictionary of words, [s], over symbols from some alphabet, Σ. Most operations require that s be an instance of the Ord class.

The tree is designed specifically to make it easy to query for (sub)anagrams, i.e. words w and w' where sort w and sort w' are equivalent. The time to do most of these operations are mainly dependent on the size of the tree (t) the alphabet (|Σ|) the output (m), and not the number of words stored (n).

This module is intended to be imported qualified. That is, please import it as follows:

import qualified Anatree as Anatree

Each set of anagrams are stored in a bucket via a Set. Hence, no more than maxBound::Int anagram "collisions" can be stored at once.

Tree Type

data Tree s

Each node in the tree contains a Set s (/s is most often Char) with the words that are each others' anagrams. Each Node contains a binary choice of whether a certain character exists in the normalized (sorted) word.

Constructors

Leaf (Set [s])

Node (Set [s]) s (Tree s) (Tree s)

type Size = Int

Construction

empty :: Tree s

O(1) Creates an empty anagram tree.

singleton :: Ord s => [s] -> Tree s

O(|w| log(|w|)) Creates an anagram tree with a single word.

fromList :: Ord s => [[s]] -> Tree s

Repeated use of insert into an empty anagram tree.

fromSet :: Ord s => Set [s] -> Tree s

Repeated use of insert into an empty anagram tree.

Insertion

insert :: Ord s => [s] -> Tree s -> Tree s

O(|w| log |w| + |Σ|) Add word to the set.

Deletion

delete :: Ord s => [s] -> Tree s -> Tree s

O(|w| log |w| + |Σ|) Remove word from the set.

Queries

anagrams :: Ord s => [s] -> Tree s -> Set [s]

O(|w| log |w| + |Σ|) The anagrams that exists in the set.

member :: Ord s => [s] -> Tree s -> Bool

O(|w| log |w| + |Σ|) Is the word in the set?

notMember :: Ord s => [s] -> Tree s -> Bool

O(|w| log |w| + |Σ|) Is the word not in the set?

subanagrams :: Ord s => [s] -> Tree s -> Set [s]

O(|w| log |w| + m(|Σ| |w|)) The subanagrams that exists in the set.

keys :: Ord s => Int -> Tree s -> Set [s]

O(n²) The words of a certain length that are not anagrams of each other.

null :: Tree s -> Bool

O(1) Is this the empty set?

size :: Tree s -> Int

O(t) Number of elements in the set.

treeSize :: Tree s -> Int

O(t) Size of the tree itself.

Combine

union :: Ord s => Tree s -> Tree s -> Tree s

The union of two sets, preferring the first set when equal elements are encountered.

unions :: (Foldable f, Ord s) => f (Tree s) -> Tree s

The union over a Foldable structure, i.e. foldl union empty.

intersection :: Ord s => Tree s -> Tree s -> Tree s

The intersection of two sets. Elements of the result come from the first set.

intersections :: (Foldable f, Ord s) => f (Tree s) -> Tree s

The intersection over a Foldable structure, i.e. foldl1 intersection.

Folds

foldr :: ([s] -> t -> t) -> t -> Tree s -> t

Fold the words in the set using a right-associative binary operator.

foldl :: (t -> [s] -> t) -> t -> Tree s -> t

Fold the words in the set using a left-associative binary operator.

foldMap :: Monoid m => ([s] -> m) -> Tree s -> m

Fold mapped words in the set.

Conversion

List

elems :: Tree s -> [[s]]

An alias of toList.

toList :: Tree s -> [[s]]

Convert the set to a list of elements.

toAscList :: Ord s => Tree s -> [[s]]

Convert the set into an ascending list of elements.

toDescList :: Ord s => Tree s -> [[s]]

Convert the set into a descending list of elements.

Set

toSet :: Ord s => Tree s -> Set [s]

Convert the set of words to a Set.